// src/app/modules/hcm/shared/models/salaries/rate-exchange-xtype-salary.ts
export class RateExchangexTypeSalary{
    idTypeSalary: number = -1;
    typeSalary: string = "";
    idCurrency: number = -1;
    abbreviation: string = "";
    currency: string = "";
    symbol: string = "";
    conversionFactor: number = -1;
    dateFactor:string = '';
    idCurrencyDestination: number = -1;
    abbreviationDestination: string = "";
    currencyDestination: string = "";
    symbolDestination: string = "";
}